# z-reader-setting/src/main/java/com/zreader/setting/model/CategoryDTO.java
package com.zreader.setting.model;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

/**
 * 书籍分类请求 DTO
 *
 * @author zy
 * @date 2025/11/12
 */
public class CategoryDTO {

    /**
     * 分类名称
     */
    @NotBlank(message = "分类名称不能为空")
    @Size(min = 1, max = 20, message = "分类名称长度必须在1-20个字符之间")
    private String categoryName;

    public String getCategoryName() {
        return categoryName;
    }

    public void setCategoryName(String categoryName) {
        this.categoryName = categoryName;
    }
}